requirements.txt
Django==1.8.1
django-axes==1.3.9
django-cors-headers==1.0.0
django-debug-toolbar==1.3.0
django-extensions==1.5.1
django-redis==3.8.3
django-suit==0.2.13
djangorestframework==3.1.0
djangorestframework-gis==0.8.2
docutils==0.12
Markdown==2.6.1
Pillow==2.8.1
postgres==2.1.2
psycopg2==2.6
pyt==0.7.6
pytz==2014.10
redis==2.10.3
six==1.5.2
sqlparse==0.1.14
suit==2.0.2
